refactor: move primitives + shared behind @design-system seam
Consolidate the in-tree design-system into one location and expose it
through a single import seam so a downstream productionizer can swap
the component layer (or the token values) without forking the repo.

- git mv primitives (Button, Chip, EditableField) →
  src/design-system/primitives/
- git mv shared-composed (Card, StatusBadge, ErrorState,
  ErrorBoundary[.test], UpdateBanner) → src/design-system/shared/
- git mv styles (theme.css, tokens.css) → src/design-system/styles/
- Add src/design-system/index.ts barrel; feature code now imports via
  `@design-system`, never deep paths
- Add @design-system Vite alias (build + vitest) and tsconfig paths;
  repoint @design-tokens default to the moved tokens.css; styles.css
  imports moved theme.css
- Update all feature consumers (App.tsx, Result.tsx, ContactCard,
  FeedbackControl, JdInput, JdMatch, ReconstructedRole, RewriteButton)
- Neutralize the default token palette (navy/amber/cream → generic
  slate + blue accent), token NAMES unchanged, light + dark modes
- Update eslint.config.js architecture guard to cover
  src/design-system and repoint Button allowlist path
- Update scripts/hooks/style_guard.sh scope + BUTTON_PRIMITIVE path
- Update README.md Theming section (paths, generic-palette note, new
  "Swapping the component layer" subsection with component API table)
- Update CLAUDE.md 3-tier paths + @design-system seam description

Resolves #105
