+++ src/components/Result.tsx
@@ -5,9 +5,7 @@ import type { CascadeResult } from "../lib/heuristics/types.ts";
 import type { AnonymousAtsScore } from "../lib/score/score.ts";
 import { getScoreLabel, getScoreTier } from "../lib/score/score.ts";
 import type { EditableParse } from "../hooks/useEditableParse.ts";
-import { Card } from "./shared/Card.tsx";
-import { StatusBadge } from "./shared/StatusBadge.tsx";
-import { Button } from "./ui/Button.tsx";
+import { Card, StatusBadge, Button } from "@design-system";
 import { FeedbackControl } from "./features/FeedbackControl.tsx";
 import { ReconstructedResume } from "./features/ReconstructedResume.tsx";
 import { AtsScoreReadout } from "./features/AtsScoreReadout.tsx";
